+++ saas-ui/pro/src/filters/provider.tsx
@@ -53,7 +53,7 @@ export const FiltersProvider: React.FC<FiltersProviderProps> = (props) => {
   }
 
   const getActiveFilters = (): Filter[] =>
-    [...activeFilterMap].map(([key, filter]) => ({
+    Array.from(activeFilterMap.entries()).map(([key, filter]) => ({
       key,
       ...filter,
     }))