+++ saas-ui/pro/src/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './use-search-query'
-export * from './use-active-path'
 export * from './use-script'