+++ saas-ui/pro/src/sidebar/sidebar.tsx
@@ -39,12 +39,12 @@ export {
   MenuDivider as SidebarMenuDivider,
 } from '@chakra-ui/menu'
 
-import { ChevronDownIcon, ChevronRightIcon, HamburgerIcon } from './icons'
+import { ChevronDownIcon, ChevronRightIcon, HamburgerIcon } from '../icons'
 
 import { SidebarProvider, useSidebarContext } from './use-sidebar'
 import { ResizeHandle, ResizeHandler, useResize } from '../resize'
 
-import { useActivePath } from '..'
+import { useActivePath } from '@saas-ui/router'
 
 export interface SidebarProps
   extends Omit<HTMLMotionProps<'div'>, 'color' | 'transition'>,