+++ packages/app-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @app/config
 
+## 0.0.12
+
+### Patch Changes
+
+- @ui/core@0.0.21
+
+## 0.0.11
+
+### Patch Changes
+
+- @ui/core@0.0.20
+
 ## 0.0.10
 
 ### Patch Changes