+++ packages/app-graphql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @app/graphql
 
+## 0.1.3
+
+### Patch Changes
+
+- @app/config@0.0.12
+
+## 0.1.2
+
+### Patch Changes
+
+- @app/config@0.0.11
+
 ## 0.1.1
 
 ### Patch Changes