+++ packages/app-mock-graphql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @app/mock-graphql
 
+## 0.3.2
+
+### Patch Changes
+
+- @app/graphql@0.1.3
+
+## 0.3.1
+
+### Patch Changes
+
+- @app/graphql@0.1.2
+
 ## 0.3.0
 
 ### Minor Changes